+++ python/old/straw.cpp
